src/dict_cost.py

def get_cost(costs:dict = {}, items:list = [], tax:int = 0.69) -> float:
    """
    This function takes a dictionary as input and returns a list of costs based on existing items
    If the input dictionary is empty, it returns a list with a single value of 0.0.
    Otherwise, it returns a list with the value of the 'cost' key in the input dictionary.
    :param costs: A dictionary containing items and their corresponding costs.
    :param items: A list of items for which the cost needs to be calculated.
    :param tax: A tax rate to be applied to the cost.
    :return: The total cost including the tax
    """

    if costs.__len__() == 0 or items.__len__()== 0:
        return float(0)
    else:
        cost = [ value for key, value in costs.items() if key in items ]
        cost = sum(cost)*(1+tax)
        return round(cost, 2)

if __name__ == "__main__":
    """
    This is the main function that demonstrates the usage of the get_cost function.
    It creates a dictionary 'costs' with a 'cost' key and a value of 100.
    It then calls the get_cost function with the 'cost' dictionary and prints the result.
    """
    costs = {'socks': 5, 'shoes': 60, 'sweater': 30}
    items = ['socks', 'shoes']
    tax = 0.09
    print(get_cost(costs, items, tax))